Rename util/conv to util/compare. Update godoc for IsNameSame()
